chore: move pages to src
