+++ __pages__/index.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from "@storybook/react"
 
-import Home, { HomeProps } from "pages/index"
+import Home, { HomeProps } from "src/pages/index"
 
 import { generateBlogPostDocument } from "lib/prismic/util"
