+++ __pages__/project.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from "@storybook/react"
 
-import Project, { ProjectProps } from "pages/project/[slug]"
+import Project, { ProjectProps } from "src/pages/project/[slug]"
 
 import { generateProjectDocument } from "lib/prismic/util"
