+++ src/pages/project/[slug].tsx
